Split ContextManagerBuilder for build one time only and put renew into coordinator
